refactor(StationViewEventDelegator) : move StationViewEventDelegator to eventdelegator folder
