+++ src/eventDelegators/StationViewEventDelegator/index.js
@@ -1,0 +1,38 @@
+export default class StationViewEventDelegator {
+  constructor(element, stationView, subwayMapViewModel) {
+    this.stationView = stationView;
+    this.subwayMapViewModel = subwayMapViewModel;
+    element.addEventListener('click', this.onClick.bind(this));
+  }
+
+  onClick(event) {
+    let dataSet = event.target.dataset;
+
+    if (dataSet.purpose) {
+      this[dataSet.purpose](dataSet);
+    }
+  }
+
+  stationManager() {
+    this.stationView.resetManagerContainer();
+    this.stationView.renderStationManager();
+  }
+
+  addStation() {
+    const stationId = document.getElementById('#station-name-input').value;
+    this.subwayMapViewModel.addStation(stationId);
+
+    this.stationView.resetStationTable();
+    this.stationView.renderStationTable(
+      Object.entries(this.subwayMapViewModel.getStations()),
+    );
+  }
+
+  deleteStation(dataSet) {
+    this.subwayMapViewModel.deleteStation(dataSet.stationid);
+    this.stationView.resetStationTable();
+    this.stationView.renderStationTable(
+      Object.entries(this.subwayMapViewModel.getStations()),
+    );
+  }
+}
