+++ src/eventDelegators/index.js
@@ -1,0 +1,3 @@
+import StationViewEventDelegator from './StationViewEventDelegator';
+
+export { StationViewEventDelegator };
